LayoutTests/fast/dom/shadow/shadow-root-new-expected.txt

This test checks the object 'new WebKitShadowRoot(host)' and internals.shadowRoot(div) are the same.

PASS root1 is root2
PASS successfullyParsed is true

TEST COMPLETE
